Move decorator to layer.py
