+++ batchflow/models/tf/layers/pooling.py
@@ -4,8 +4,7 @@ import numpy as np
 import tensorflow as tf
 import tensorflow.keras.layers as K #pylint: disable=import-error
 
-from .layer import Layer
-from .utils import add_as_function
+from .layer import Layer, add_as_function
